+++ users/RELEASES.md
@@ -90,3 +90,13 @@ Based on:
 - [csharp v3.0.1] users
 ### Releases
 - [NuGet v3.0.1] https://www.nuget.org/packages/WingspanUsers/3.0.1 - users
+
+## 2024-02-05 01:26:41
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.166.0 (2.248.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.0.0] users
+### Releases
+- [NuGet v4.0.0] https://www.nuget.org/packages/WingspanUsers/4.0.0 - users